fix(wikia): add areaAttack from wikia data
areaAttack information missing from various weapons.  This is relevant
to all weapons with a radial damage value such as the Kulstar, Angstrum,
etc.

When looking into this information it was noticed that the damagePerShot
array includes the damage information for the radial Damage, but does
not include relevant information to help distinguish radial damage from
primary damage.  This is relevant as weapons with Radial Damage have
double-gains from all damage mods (adding Corrosion is applied to the
Primary damage and again via Radial Damage). In effect, if a weapon
supports radial damage, then the damagePerShot array is unusable for
calculating DPS of the weapon.